+++ lib/tus/errors.rb
@@ -1,3 +1,5 @@
+# frozen-string-literal: true
+
 module Tus
   Error           = Class.new(StandardError)
   NotFound        = Class.new(Error)