+++ lib/tus/storage/filesystem.rb
@@ -1,3 +1,4 @@
+# frozen-string-literal: true
 require "tus/errors"
 
 require "pathname"