+++ lib/tus/storage/s3.rb
@@ -1,3 +1,4 @@
+# frozen-string-literal: true
 require "aws-sdk"
 
 require "tus/info"