+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Deprecate `Tus::Server::Goliath` in favour of [goliath-rack_proxy](https://github.com/janko-m/goliath-rack_proxy) (@janko-m)
+
 * Reduce string allocations in MRI 2.3+ with `frozen-string-literal: true` magic comments (@janko-m)
 
 ## 1.0.0 (2017-07-17)