+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## HEAD
+## 1.1.0 (2017-07-23)
 
 * Ignore retryable networking errors in `Tus::Storage::S3#patch_file` for resiliency (@janko-m)